// src/models/machines.ts
/**
 * Ravimoshark Contract API
 *
 * OpenAPI spec version: 1.0.0
 * Contact: dev@ravimoshark.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * Client's machines associated with a contract.
 */
export interface Machines {
    id?: number;
    piece?: string;
    numSerie?: string;
    refContract?: string;
}